Revert "Fix compilation with MSVC: arm64/disassembler/decode_scratchpad.c"
This reverts commit cb7c2d3bb716a2e96cf7f127bbf71992a4861907.

+++ arch/arm64/disassembler/decode_scratchpad.c
@@ -4625,7 +4625,7 @@ int decode_scratchpad(context* ctx, Instruction* instr)
 		case ENC_BFI_BFM_32M_BITFIELD:
 		case ENC_SBFIZ_SBFM_32M_BITFIELD:
 		case ENC_UBFIZ_UBFM_32M_BITFIELD:
-			lsb = -(int64_t)IMMR % 32;
+			lsb = -IMMR % 32;
 			width = IMMS + 1;
 			break;
 		default:
